Refactor and improv of multinode tests
This commit changes the integration test, by splitting it from a single
process into multiple process executing them on different emulated
machines trough docker-compose. It also refactor the tests directory to
better reflect the multinode nature of the tests
