+++ .github/workflows/ci-tests.yaml
@@ -55,18 +55,14 @@ jobs:
               --gtest_break_on_failure            \
               --gtest_print_time=1
 
-          echo "Run CAPIO integration tests"
-            docker run --rm                     \
-            --env CAPIO_DIR=/tmp                \
-            --env CAPIO_LOG_LEVEL=-1            \
-            --env LD_PRELOAD=libcapio_posix.so  \
-            --name capio-docker                 \
-            alphaunito/capio:latest             \
-            capio_integration_tests             \
-            --gtest_break_on_failure            \
-            --gtest_print_time=1
-          
-          docker compose -f tests/docker-compose.yml up --abort-on-container-exit
+      - name: "Run backend communication tests"
+        working-directory: ${{ github.workspace }}
+        run: docker compose -f tests/multinode/backend/docker-compose.yml up --abort-on-container-exit
+
+      - name: "Run integration tests"
+        working-directory: ${{ github.workspace }}
+        run: docker compose -f tests/multinode/integration/docker-compose.yml up --abort-on-container-exit
+
 
   format-check:
     name: "Check ${{ matrix.path }} clang-format conformance"
@@ -159,6 +155,12 @@ jobs:
         run: |
           export LD_LIBRARY_PATH="/usr/local/lib:${LD_LIBRARY_PATH}"
           
+          echo "Run CAPIO syscall Unit tests"
+          LD_PRELOAD=libcapio_posix.so \
+          capio_syscall_unit_tests \
+            --gtest_break_on_failure \
+            --gtest_print_time=1
+          
           echo "Run CAPIO SERVER unit tests"
           capio_server_unit_tests \
             --gtest_break_on_failure \
@@ -167,25 +169,8 @@ jobs:
           echo "Run CAPIO POSIX Unit tests"
           capio_posix_unit_tests \
             --gtest_break_on_failure \
-            --gtest_print_time=1
-          
-          echo "Run CAPIO SERVER unit tests"
-          capio_server_unit_tests \
-            --gtest_break_on_failure \
-            --gtest_print_time=1
-          
-          echo "Run CAPIO syscall Unit tests"
-          LD_PRELOAD=libcapio_posix.so \
-          capio_syscall_unit_tests \
-            --gtest_break_on_failure \
-            --gtest_print_time=1
-          
-          echo "Run CAPIO integration tests"
-          rm -rf /dev/shm/CAPIO*
-          LD_PRELOAD=libcapio_posix.so \
-          capio_integration_tests \
-            --gtest_break_on_failure \
-            --gtest_print_time=1
+            --gtest_print_time=1 
+
       - name: "Show client logs on failure"
         if: ${{ always() && steps.run-tests.outcome == 'failure' && matrix.build_type == 'Debug' }}
         run: tail -v -n +1 capio_logs/posix/$(hostname)/posix_thread_*.log
