+++ tests/CMakeLists.txt
@@ -25,5 +25,5 @@ include(GoogleTest)
 add_subdirectory(unit/posix)
 add_subdirectory(unit/server)
 add_subdirectory(unit/syscall)
-add_subdirectory(unit/backend)
-add_subdirectory(integration)
+add_subdirectory(multinode/integration)
+add_subdirectory(multinode/backend)
